backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sqlalchemy
psycopg2-binary  
supabase
celery
redis
aiohttp 
python-dotenv 
openai